Incremental commit, won't work if you pull it.

+++ newsroom-api/src/services/RoleService.ts
@@ -1,44 +1,50 @@
 import * as express from "express";
 import { getManager } from "typeorm";
-import { DELETE, Errors, GET, Path, PathParam, POST, PreProcessor, PUT } from "typescript-rest";
+import { Context, DELETE, GET, Path, PathParam, POST, PreProcessor, PUT, ServiceContext } from "typescript-rest";
 import { IsInt, Tags } from "typescript-rest-swagger";
-import { NotFoundError } from "typescript-rest/dist/server/model/errors";
 
+import { NRPermission } from "../entity/NRPermission";
 import { NRRole } from "../entity/NRRole";
+import { NRType } from "../entity/NRType";
 import { NRUser } from "../entity/NRUser";
+import { common } from "./Common";
 import { validators } from "./Validators";
 
-/**
- * Provides API services for roles.
- */
+// Provides API services for roles.
 @Path("/api/roles")
 @Tags("Roles")
 export class RoleService {
-    /**
-     * Used to interact with any specified type in the database.
-     */
+    // Context manager to grab injected user from the request.
+    @Context
+    private context: ServiceContext;
+
+    // Database interactions managers.
     private roleRepository = getManager().getRepository(NRRole);
+    private permRepository = getManager().getRepository(NRPermission);
 
     /**
-     * Create a new entry in the 'role' table with the specified
-     * information.
+     * Create a new role.
      *
-     * Returns 400 if:
-     *      -
-     *
-     * Returns 404 if:
-     *      - Bad parameters.
+     * Returns:
+     *      - NRRole
+     *      - BadRequestError (400)
+     *          - If form submission data is invalid.
      */
     @POST
     @PreProcessor(validators.createRoleValidator)
     public async createRole(role: NRRole): Promise<NRRole> {
-        // Form data already validated above.
+        const sessionUser = common.getUserFromContext(this.context);
+        await common.checkAdmin(sessionUser);
+
+        // Form data already validated.
         return await this.roleRepository.save(role);
     }
 
     /**
-     * Get all roles that exist in the 'role' table under
-     * the configured connection.
+     * Get all existing roles.
+     *
+     * Returns:
+     *      - NRRole[]
      */
     @GET
     public async getAllRoles(): Promise<NRRole[]> {
@@ -46,46 +52,37 @@ export class RoleService {
     }
 
     /**
-     * Get a specific role from the 'role' table based on passed
-     * role id.
+     * Get a specific role by ID.
      *
-     * Returns 404 if:
-     *      - Role now found.
+     * Returns:
+     *      - NRRole
+     *      - NotFoundError (404)
+     *          - If role not found.
      */
     @GET
-    @Path(":/id")
-    public async getRole(@IsInt @PathParam("id") roleId: number): Promise<NRRole> {
-        try {
-            return await this.roleRepository.findOneOrFail(roleId);
-        } catch (err) {
-            console.error("Error getting Role:", err);
-            throw new NotFoundError("A Role with the given ID was not found.");
-        }
+    @Path(":/rid")
+    public async getRole(@IsInt @PathParam("rid") rid: number): Promise<NRRole> {
+        return await common.getRole(rid, this.roleRepository);
     }
 
     /**
-     * Update an entry in the 'role' table with the specified
-     * information.
+     * Update a given role.
      *
-     * Return 400 if:
-     *      - Role property types incorrect.
-     *
-     * Return 404 if:
-     *      - Role id field is missing or not found.
+     * Returns:
+     *      - NRRole
+     *      - BadRequestError (400)
+     *          - If proeprties incorrect.
+     *      - NotFoundError (404)
+     *          - If role not found.
      */
     @PUT
-    @Path("/:id")
+    @Path("/:rid")
     @PreProcessor(validators.updateRoleValidator)
-    public async updateRole(@IsInt @PathParam("id") roleId: number,
+    public async updateRole(@IsInt @PathParam("rid") rid: number,
                             role: NRRole): Promise<NRRole> {
-        let currRole: NRRole;
-
-        try {
-            currRole = await this.roleRepository.findOneOrFail(roleId);
-        } catch (err) {
-            console.error("Error getting Role:", err);
-            throw new NotFoundError("A Role with the given ID was not found.");
-        }
+        const sessionUser = common.getUserFromContext(this.context);
+        const currRole = await common.getRole(rid, this.roleRepository);
+        await common.checkWritePermissions(sessionUser, NRType.ROLE_KEY, rid);
 
         // Update current stored name if given one.
         if (role.name) {
@@ -98,21 +95,66 @@ export class RoleService {
     /**
      * Delete a role.
      *
-     * Returns 404 if:
-     *      - Role id not found.
+     * Returns:
+     *      - NotFoundError (404)
+     *          - If role not found.
      */
     @DELETE
-    @Path("/:id")
-    public async deleteRole(@IsInt @PathParam("id") roleId: number) {
-        let currRole: NRRole;
-
-        try {
-            currRole = await this.roleRepository.findOneOrFail(roleId);
-        } catch (err) {
-            console.error("Error getting Role:", err);
-            throw new NotFoundError("A Role with the given ID was not found.");
-        }
+    @Path("/:rid")
+    public async deleteRole(@IsInt @PathParam("rid") rid: number) {
+        const sessionUser = common.getUserFromContext(this.context);
+        const currRole = await common.getRole(rid, this.roleRepository);
+        await common.checkWritePermissions(sessionUser, NRType.ROLE_KEY, rid);
 
         await this.roleRepository.remove(currRole);
     }
+
+    /**
+     * Add a permission to a role.
+     *
+     * Returns:
+     *      - NRRole
+     *      - NotFoundError (404)
+     *          - If role not found.
+     */
+    @PUT
+    @Path("/:rid/add/:pid")
+    public async addPermission(@IsInt @PathParam("rid") rid: number,
+                               @IsInt @PathParam("pid") pid: number): Promise<NRRole> {
+        const sessionUser = common.getUserFromContext(this.context);
+        const currRole = await common.getRole(rid, this.roleRepository);
+        await common.checkWritePermissions(sessionUser, NRType.ROLE_KEY, rid);
+
+        // Check existence.
+        const newPermission = await common.getPermission(pid, this.permRepository);
+
+        currRole.permissions.push(newPermission);
+        return await this.roleRepository.save(currRole);
+    }
+
+   /**
+    * Remove a permission to a role.
+    *
+    * Returns:
+    *      - NRRole
+    *      - NotFoundError (404)
+    *          - If role not found.
+    */
+    @PUT
+    @Path("/:rid/remove/:pid")
+    public async removePermission(@IsInt @PathParam("rid") rid: number,
+                                  @IsInt @PathParam("pid") pid: number): Promise<NRRole> {
+        const sessionUser = common.getUserFromContext(this.context);
+        const currRole = await common.getRole(rid, this.roleRepository);
+        await common.checkWritePermissions(sessionUser, NRType.ROLE_KEY, rid);
+
+        // Check existence.
+        const newPermission = await common.getPermission(pid, this.permRepository);
+
+        // Remove relationship and save.
+        const ind = currRole.permissions.indexOf(newPermission);
+        currRole.permissions.splice(ind, 1);
+        return await this.roleRepository.save(currRole);
+    }
+
 }
